requirements.txt
Django==4.2.8
django-allauth==0.58.2
djangorestframework==3.14.0
psycopg2-binary